Commit authorize a second time
Added functionality to prevent clicks on used squares, prevent clicks
when we have a winner, reset board and keep track of number of games
won by X or O.
Displaying name of player when logging in.
Need to work on styling and also hide signout/chgpswd when noone
signed in.